apps/infra/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
description = "Suite infra for analytics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.8.0",
    "boto3>=1.28.0",
    "celery>=5.3.0",
    "fastapi[standard]>=0.109.1",
    "pydantic>=2.7.0,<3.0.0",
    "pyiceberg[pandas,polars,pyarrow]>=0.5.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "web3>=6.0.0,<7.0.0",
]

